Minor changes to date field component
Additional changes:
- Move common function to utils/misc.js
- Adjust CSS
