+++ client/src/components/DateField.jsx
@@ -3,17 +3,15 @@ import { Typography } from '@material-ui/core';
 import { makeStyles } from '@material-ui/core/styles';
 import DatePicker from 'react-date-picker';
 import moment from 'moment';
+import { getDateParts } from '../utils/misc';
 
 const useStyles = makeStyles(theme => ({
   picker: {
     backgroundColor: '#fff',
     width: 168,
+    fontSize: 18,
   },
-  root: {
-    maxWidth: 72,
-    marginTop: theme.spacing(1.25),
-    paddingLeft: theme.spacing(1),
-  },
+  root: { marginTop: theme.spacing(1.25), paddingLeft: theme.spacing(1) },
   typography: { color: theme.palette.primary.contrastText },
 }));
 
@@ -32,16 +30,10 @@ const DateField = ({ filterID, name, value, valueObj, onChange }) => {
     }
   };
 
-  let month;
-  let day;
-  let year;
   let date;
 
-  if (value) {
-    month = value[0].substring(0, 2);
-    day = value[0].substring(3, 5);
-    year = value[0].substring(6, 10);
-
+  if (value[0]) {
+    const { month, day, year } = getDateParts(value[0]);
     date = new Date(year, month - 1, day);
   }
